+++ pxr/imaging/plugin/hdOSPRay/mesh.cpp
@@ -36,10 +36,6 @@
 #include <pxr/imaging/hd/smoothNormals.h>
 #include <pxr/imaging/pxOsd/tokens.h>
 
-#include <pxr/imaging/hdSt/drawItem.h>
-#include <pxr/imaging/hdSt/geometricShader.h>
-#include <pxr/imaging/hdSt/material.h>
-
 #include <rkcommon/math/AffineSpace.h>
 
 using namespace rkcommon::math;
@@ -179,6 +175,8 @@ HdOSPRayMesh::_UpdatePrimvarSources(HdSceneDelegate* sceneDelegate,
     HD_TRACE_FUNCTION();
     SdfPath const& id = GetId();
 
+    TF_VERIFY(sceneDelegate);
+
     // Update _primvarSourceMap, our local cache of raw primvar data.
     // This function pulls data from the scene delegate, but defers processing.
     //
@@ -204,15 +202,17 @@ HdOSPRayMesh::_UpdatePrimvarSources(HdSceneDelegate* sceneDelegate,
             // TODO: need to find a better way to identify the primvar for
             // the texture coordinates
             // texcoords
-            if ((pv.name == "Texture_uv" || pv.name == "st")
+            if ((pv.role == HdPrimvarRoleTokens->textureCoordinate
+                 || pv.name == HdOSPRayTokens->st)
                 && HdChangeTracker::IsPrimvarDirty(dirtyBits, id,
                                                    HdOSPRayTokens->st)) {
                 if (value.IsHolding<VtVec2fArray>()) {
                     _texcoords = value.Get<VtVec2fArray>();
+                    _texcoordsInterpolation = interp;
                 }
             }
 
-            if (pv.name == "displayColor"
+            if (pv.name == HdTokens->displayColor
                 && HdChangeTracker::IsPrimvarDirty(dirtyBits, id,
                                                    HdTokens->displayColor)) {
                 // Create 4 component displayColor/opacity array for OSPRay
@@ -237,7 +237,7 @@ HdOSPRayMesh::_UpdatePrimvarSources(HdSceneDelegate* sceneDelegate,
                 }
             }
 
-            if (pv.name == "displayOpacity"
+            if (pv.name == HdTokens->displayOpacity
                 && HdChangeTracker::IsPrimvarDirty(dirtyBits, id,
                                                    HdTokens->displayOpacity)) {
                 // XXX assuming displayOpacity can't exist without
@@ -388,23 +388,13 @@ HdOSPRayMesh::_PopulateOSPMesh(HdSceneDelegate* sceneDelegate,
 
         if (doRefine) {
             _ospMesh = _CreateOSPRaySubdivMesh();
-            _ospMesh.commit();
 
             HdMeshUtil meshUtil(&_topology, GetId());
-            meshUtil.ComputeQuadIndices(&_quadIndices, &_quadPrimitiveParams);
-            // Check if texcoords are provides as face varying.
-            // XXX: (This code currently only cares about _texcoords, but
-            // should be generalized to all primvars)
-            bool faceVaryingTexcoord = false;
-            HdPrimvarDescriptorVector faceVaryingPrimvars
-                   = GetPrimvarDescriptors(sceneDelegate,
-                                           HdInterpolationFaceVarying);
-            for (HdPrimvarDescriptor const& pv : faceVaryingPrimvars) {
-                if (pv.name == "Texture_uv" || pv.name == "st")
-                    faceVaryingTexcoord = true;
-            }
-
-            if (faceVaryingTexcoord) {
+            // meshUtil.ComputeQuadIndices(&_quadIndices,
+            // &_quadPrimitiveParams); Check if texcoords are provided as face
+            // varying.
+            if (_texcoordsInterpolation == HdInterpolationFaceVarying) {
+                // WARNING:  face varying currently only supported by ptex
                 TfToken buffName = HdOSPRayTokens->st;
                 VtValue buffValue = VtValue(_texcoords);
                 HdVtBufferSource buffer(buffName, buffValue);
@@ -416,8 +406,8 @@ HdOSPRayMesh::_PopulateOSPMesh(HdSceneDelegate* sceneDelegate,
                 if (success && quadPrimvar.IsHolding<VtVec2fArray>()) {
                     _texcoords = quadPrimvar.Get<VtVec2fArray>();
                 } else {
-                    std::cout << "ERROR: could not quadrangulate face-varying "
-                                 "data\n";
+                    TF_CODING_ERROR("ERROR: could not quadrangulate "
+                                    "face-varying data\n");
                 }
             }
         }
@@ -447,37 +437,24 @@ HdOSPRayMesh::_PopulateOSPMesh(HdSceneDelegate* sceneDelegate,
                                            HdOSPRayTokens->st)) {
 
         const HdRenderIndex& renderIndex = sceneDelegate->GetRenderIndex();
-        const HdOSPRayMaterial* material
-               = static_cast<const HdOSPRayMaterial*>(renderIndex.GetSprim(
-                      HdPrimTypeTokens->material, GetMaterialId()));
+        // Check if texcoords are provides as face varying.
+        // XXX: (This code currently only cares about _texcoords, but
+        // should be generalized to all primvars)
+        bool faceVaryingTexcoord = false;
+        HdPrimvarDescriptorVector faceVaryingPrimvars = GetPrimvarDescriptors(
+               sceneDelegate, HdInterpolationFaceVarying);
+        for (HdPrimvarDescriptor const& pv : faceVaryingPrimvars) {
+            if (pv.name == "Texture_uv" || pv.name == "st")
+                faceVaryingTexcoord = true;
+        }
+
+        bool useQuads = _UseQuadIndices(renderIndex, _topology);
 
         if (!_refined) {
-            bool useQuads = _UseQuadIndices(renderIndex, _topology);
-
             if (useQuads) {
-                _ospMesh = opp::Geometry("mesh");
-
                 HdMeshUtil meshUtil(&_topology, GetId());
                 meshUtil.ComputeQuadIndices(&_quadIndices,
                                             &_quadPrimitiveParams);
-
-                opp::SharedData indices = opp::SharedData(
-                       _quadIndices.cdata(), OSP_VEC4UI, _quadIndices.size());
-
-                indices.commit();
-                _ospMesh.setParam("index", indices);
-
-                // Check if texcoords are provides as face varying.
-                // XXX: (This code currently only cares about _texcoords, but
-                // should be generalized to all primvars)
-                bool faceVaryingTexcoord = false;
-                HdPrimvarDescriptorVector faceVaryingPrimvars
-                       = GetPrimvarDescriptors(sceneDelegate,
-                                               HdInterpolationFaceVarying);
-                for (HdPrimvarDescriptor const& pv : faceVaryingPrimvars) {
-                    if (pv.name == "Texture_uv" || pv.name == "st")
-                        faceVaryingTexcoord = true;
-                }
 
                 if (faceVaryingTexcoord) {
                     TfToken buffName = HdOSPRayTokens->st;
@@ -492,9 +469,8 @@ HdOSPRayMesh::_PopulateOSPMesh(HdSceneDelegate* sceneDelegate,
                     if (success && quadPrimvar.IsHolding<VtVec2fArray>()) {
                         _texcoords = quadPrimvar.Get<VtVec2fArray>();
                     } else {
-                        std::cout
-                               << "ERROR: could not quadrangulate face-varying "
-                                  "data\n";
+                        TF_CODING_ERROR("ERROR: could not quadrangulate "
+                                        "face-varying data\n");
                     }
 
                     // usd stores texcoords in face indexed -> each quad has 4
@@ -512,32 +488,10 @@ HdOSPRayMesh::_PopulateOSPMesh(HdSceneDelegate* sceneDelegate,
                     }
                     _texcoords = texcoords2;
                 }
-
-            } else { // triangles
-                _ospMesh = opp::Geometry("mesh");
-
+            } else {
                 HdMeshUtil meshUtil(&_topology, GetId());
                 meshUtil.ComputeTriangleIndices(&_triangulatedIndices,
                                                 &_trianglePrimitiveParams);
-
-                opp::SharedData indices = opp::SharedData(
-                       _triangulatedIndices.cdata(), OSP_VEC3UI,
-                       _triangulatedIndices.size());
-
-                indices.commit();
-                _ospMesh.setParam("index", indices);
-
-                // Check if texcoords are provides as face varying.
-                // XXX: (This code currently only cares about _texcoords, but
-                // should be generalized to all primvars)
-                bool faceVaryingTexcoord = false;
-                HdPrimvarDescriptorVector faceVaryingPrimvars
-                       = GetPrimvarDescriptors(sceneDelegate,
-                                               HdInterpolationFaceVarying);
-                for (HdPrimvarDescriptor const& pv : faceVaryingPrimvars) {
-                    if (pv.name == "Texture_uv" || pv.name == "st")
-                        faceVaryingTexcoord = true;
-                }
 
                 if (faceVaryingTexcoord) {
                     TfToken buffName = HdOSPRayTokens->st;
@@ -554,9 +508,8 @@ HdOSPRayMesh::_PopulateOSPMesh(HdSceneDelegate* sceneDelegate,
                         && triangulatedPrimvar.IsHolding<VtVec2fArray>()) {
                         _texcoords = triangulatedPrimvar.Get<VtVec2fArray>();
                     } else {
-                        std::cout
-                               << "ERROR: could not triangulate face-varying "
-                                  "data\n";
+                        TF_CODING_ERROR("ERROR: could not triangulate "
+                                        "face-varying data\n");
                     }
 
                     // usd stores texcoords in face indexed -> each triangle has
@@ -576,34 +529,36 @@ HdOSPRayMesh::_PopulateOSPMesh(HdSceneDelegate* sceneDelegate,
                     _texcoords = texcoords2;
                 }
             }
+
+            if ((_quadIndices.empty() && _triangulatedIndices.empty())
+                || _points.empty())
+                return; // invalid mesh
+
+            _ospMesh = _CreateOSPRayMesh(
+                   _quadIndices, _quadPrimitiveParams, _triangulatedIndices,
+                   _trianglePrimitiveParams, faceVaryingTexcoord, _texcoords,
+                   _points, _computedNormals, _colors, _refined, useQuads);
         }
 
-        opp::SharedData vertices
-               = opp::SharedData(_points.cdata(), OSP_VEC3F, _points.size());
-        vertices.commit();
-        _ospMesh.setParam("vertex.position", vertices);
+        const HdOSPRayMaterial* subsetMaterial = nullptr;
 
-        if (_computedNormals.size()) {
-            opp::SharedData normals
-                   = opp::SharedData(_computedNormals.cdata(), OSP_VEC3F,
-                                     _computedNormals.size());
-            normals.commit();
-            _ospMesh.setParam("vertex.normal", normals);
+        for (auto subset : _topology.GetGeomSubsets()) {
+            if (!TF_VERIFY(subset.type == HdGeomSubset::TypeFaceSet))
+                continue;
+
+            const HdOSPRayMaterial* material
+                   = static_cast<const HdOSPRayMaterial*>(renderIndex.GetSprim(
+                          HdPrimTypeTokens->material, subset.materialId));
+            subsetMaterial = material;
         }
 
-        if (_colors.size() > 1) {
-            opp::SharedData colors = opp::SharedData(_colors.cdata(), OSP_VEC4F,
-                                                     _colors.size());
-            colors.commit();
-            _ospMesh.setParam("vetex.color", colors);
-        }
-
-        if (_texcoords.size() > 1) {
-            opp::SharedData texcoords = opp::SharedData(
-                   _texcoords.cdata(), OSP_VEC2F, _texcoords.size());
-            texcoords.commit();
-            _ospMesh.setParam("vertex.texcoord", texcoords);
-        }
+        const HdOSPRayMaterial* material = nullptr;
+        if (subsetMaterial)
+            material = subsetMaterial;
+        else
+            material
+                   = static_cast<const HdOSPRayMaterial*>(renderIndex.GetSprim(
+                          HdPrimTypeTokens->material, GetMaterialId()));
 
         opp::Material ospMaterial;
 
@@ -682,11 +637,14 @@ HdOSPRayMesh::_PopulateOSPMesh(HdSceneDelegate* sceneDelegate,
                          vec3f(xfmf[12], xfmf[13], xfmf[14]));
             instance.setParam("xfm", xfm);
             instance.commit();
-            group.setParam("geometry", opp::CopiedData(*_geometricModel));
+            if (_geomSubsetModels.size()) {
+                group.setParam("geometry", opp::CopiedData(_geomSubsetModels));
+            } else {
+                group.setParam("geometry", opp::CopiedData(*_geometricModel));
+            }
             group.commit();
             _ospInstances.push_back(instance);
         }
-
         renderParam->UpdateModelVersion();
     }
     if (!_populated) {
@@ -733,6 +691,64 @@ HdOSPRayMesh::_UpdateDrawItemGeometricShader(HdSceneDelegate* sceneDelegate,
     // See HdSt_GeometricShader::BindResources()
     // The batches need to be validated and rebuilt if necessary.
     renderIndex.GetChangeTracker().MarkBatchesDirty();
+}
+
+opp::Geometry
+HdOSPRayMesh::_CreateOSPRayMesh(
+       const VtVec4iArray& quadIndices, const VtVec2iArray& quadPrimitiveParams,
+       const VtVec3iArray& triangulatedIndices,
+       const VtIntArray& trianglePrimitiveParams, bool faceVaryingTexcoord,
+       const VtVec2fArray& texcoords, const VtVec3fArray& points,
+       const VtVec3fArray& normals, const VtVec4fArray& colors, bool refined,
+       bool useQuads)
+{
+    opp::Geometry ospMesh = opp::Geometry("mesh");
+    if (!_refined) {
+        if (useQuads) {
+            opp::SharedData indices = opp::SharedData(
+                   quadIndices.cdata(), OSP_VEC4UI, quadIndices.size());
+
+            indices.commit();
+            ospMesh.setParam("index", indices);
+
+        } else { // triangles
+            opp::SharedData indices
+                   = opp::SharedData(triangulatedIndices.cdata(), OSP_VEC3UI,
+                                     triangulatedIndices.size());
+
+            indices.commit();
+            ospMesh.setParam("index", indices);
+        }
+    }
+
+    opp::SharedData verticesData
+           = opp::SharedData(points.cdata(), OSP_VEC3F, points.size());
+    verticesData.commit();
+    ospMesh.setParam("vertex.position", verticesData);
+
+    if (normals.size()) {
+        opp::SharedData normalsData
+               = opp::SharedData(normals.cdata(), OSP_VEC3F, normals.size());
+        normalsData.commit();
+        ospMesh.setParam("vertex.normal", normalsData);
+    }
+
+    if (colors.size() > 1) {
+        opp::SharedData colorsData
+               = opp::SharedData(colors.cdata(), OSP_VEC4F, colors.size());
+        colorsData.commit();
+        ospMesh.setParam("vetex.color", colorsData);
+    }
+
+    if (texcoords.size() > 1) {
+        opp::SharedData texcoordsData = opp::SharedData(
+               texcoords.cdata(), OSP_VEC2F, texcoords.size());
+        texcoordsData.commit();
+        ospMesh.setParam("vertex.texcoord", texcoordsData);
+    }
+
+    ospMesh.commit();
+    return ospMesh;
 }
 
 opp::Geometry